feat: add supplier repository

+++ Entities/Supplier.cs
@@ -11,7 +11,7 @@ namespace Entities
         public int IdProveedor { get; set; }   // PK autoincremental
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Cuit { get; set; }
+        public long Cuit { get; set; }
         public int Telefono { get; set; }
         public string Email { get; set; }
         public bool Estado { get; set; }     // true = activo, false = inactivo
